修改删除图片的接口

+++ routes/photo.js
@@ -72,13 +72,21 @@ router.post('/edit', function(req, res, next) {
 
 //把上传了的图片删除
 router.post('/delete_pic', function(req, res, next) {
-    let pic_name = req.body.pic_name;  //数据库中的图片名称
-    let id =  req.body.id;    //相片id
-    var url = path.join(__dirname,"../public/static/img/" + pic_name);
-    fs.unlink(url, (err) => {
-        if (err) throw err;
-        res.send("1")
-    });
+    try {
+        let pic_name = req.body.pic_name;  //数据库中的图片名称
+        if(pic_name){
+            var url = path.join(__dirname,"../public/static/img/" + pic_name);
+            fs.unlink(url, (err) => {
+                if (err) throw err;
+                res.send("1")
+            });
+        }else{
+            res.send("no")
+        }   
+    } catch (error) {
+        res.send("0");
+    }
+   
 });
 
 module.exports = router;
